[refactor] test_fullwrite_command로 test code 이동
